Segregated various components and extracted properties of synonyms
